support fp16 inference

+++ detic/modeling/backbone/swintransformer.py
@@ -151,7 +151,7 @@ class WindowAttention(nn.Module):
 
         attn = self.attn_drop(attn)
 
-        x = (attn @ v).transpose(1, 2).reshape(B_, N, C)
+        x = (attn.type_as(v) @ v).transpose(1, 2).reshape(B_, N, C)  # TODO: support fp16 better
         x = self.proj(x)
         x = self.proj_drop(x)
         return x
